% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateLoad.R
\name{calculateEpletLoad}
\alias{calculateEpletLoad}
\title{Calculate Eplet Load Between Donor and Recipient Genotypes}
\usage{
calculateEpletLoad(
  recipient_geno,
  donor_geno,
  loci = NULL,
  evidence_level = c("A1", "A2")
)
}
\arguments{
\item{recipient_geno}{An `hla_genotype` object for the recipient.}

\item{donor_geno}{An `hla_genotype` object for the donor.}

\item{loci}{Specific HLA loci to calculate the load across, NULL will calculate
across all loci.}

\item{evidence_level}{A character vector specifying the evidence levels for
eplets to be included in the analysis. Defaults to `c("A1", "A2")`.}
}
\value{
An integer representing the total eplet load.
}
\description{
This function calculates the total eplet load between a donor and a recipient.
It iterates over all specified HLA loci, comparing each recipient allele to each
donor allele in an exhaustive manner.
}
\examples{
# Create dummy recipient and donor genotypes
recipient <- data.frame(
  A_1 = "A*01:01", A_2 = "A*02:01",
  B_1 = "B*07:02", B_2 = "B*08:01"
)
donor <- data.frame(
  A_1 = "A*03:01", A_2 = "A*24:02",
  B_1 = "B*44:02", B_2 = "B*51:01"
)

# Convert to hla_genotype objects
rgeno <- hlaGeno(recipient)
dgeno <- hlaGeno(donor)

# Calculate eplet load
calculateEpletLoad(rgeno, dgeno)

}
