% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMatchR_eplets.R
\docType{data}
\name{deepMatchR_eplets}
\alias{deepMatchR_eplets}
\title{Eplet information extracted from [HLA Eplet Registry](https://www.epregistry.com.br/).
Eplets were filtered to retain evidence classes A1, A2, B and D).}
\description{
Eplet information extracted from [HLA Eplet Registry](https://www.epregistry.com.br/).
Eplets were filtered to retain evidence classes A1, A2, B and D).
}
