% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deepMatchR_wmda_serology}
\alias{deepMatchR_wmda_serology}
\title{WMDA DNA-to-Serology Mapping}
\format{
A data.table with the following columns:
\describe{
  \item{locus}{`character`. HLA locus with asterisk (e.g., `"A*"`, `"B*"`, `"DRB1*"`).}
  \item{allele_2f}{`character`. Two-field allele designation (e.g., `"01:01"`, `"07:02"`).}
  \item{serology}{`character`. Serological antigen number (e.g., `"1"`, `"7"`).}
}
}
\source{
WMDA nomenclature files from IMGT/HLA GitHub repository
  (\url{https://github.com/ANHIG/IMGTHLA/tree/Latest/wmda}).
}
\usage{
data(deepMatchR_wmda_serology)
}
\description{
A lookup table mapping HLA alleles to their serological equivalents based on
WMDA (World Marrow Donor Association) nomenclature. Used internally by
\code{\link{toSerology}} for allele-to-serology conversion.
}
\details{
The mapping prioritizes serology assignments in this order:
\enumerate{
  \item Unambiguous assignments
  \item Possible assignments
  \item Assumed assignments
  \item Expert assignments
}
}
\seealso{
\code{\link{toSerology}}, \code{\link{updateWmdaData}},
  \code{\link{deepMatchR_wmda_splits}}, \code{\link{deepMatchR_wmda_pgroups}}
}
\keyword{HLA}
\keyword{WMDA}
\keyword{datasets}
\keyword{serology}
