% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basilisk.R
\name{deepmatchrEnv}
\alias{deepmatchrEnv}
\title{Return a Basilisk environment tailored to the current platform}
\usage{
deepmatchrEnv(platform = c("auto", "linux", "macos"))
}
\arguments{
\item{platform}{"auto", "linux", or "macos"}
}
\description{
Return a Basilisk environment tailored to the current platform
}
