% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePeptideBindingLoad.R
\name{.deriveMismatchedPeptides}
\alias{.deriveMismatchedPeptides}
\title{Derive mismatched peptides from sequence comparison}
\usage{
.deriveMismatchedPeptides(donor_alleles, recipient_alleles, peptide_length)
}
\description{
Derive mismatched peptides from sequence comparison
}
\keyword{internal}
