% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePeptideBindingLoad.R
\name{.extractAlleles}
\alias{.extractAlleles}
\title{Extract allele names from various input types}
\usage{
.extractAlleles(x)
}
\description{
Extract allele names from various input types
}
\keyword{internal}
