% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePeptideBindingLoad.R
\name{.getPeptides}
\alias{.getPeptides}
\title{Get peptides from donor input}
\usage{
.getPeptides(donor, recipient, peptide_length)
}
\description{
Get peptides from donor input
}
\keyword{internal}
