% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePeptideBindingLoad.R
\name{.predictBinding}
\alias{.predictBinding}
\title{Predict binding using selected backend}
\usage{
.predictBinding(peptides, alleles, backend, backend_path, peptide_length)
}
\description{
Predict binding using selected backend
}
\keyword{internal}
