% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePeptideBindingLoad.R
\name{.predictBindingMHCflurry}
\alias{.predictBindingMHCflurry}
\title{MHCflurry-based binding prediction}
\usage{
.predictBindingMHCflurry(peptides, alleles)
}
\description{
MHCflurry-based binding prediction
}
\keyword{internal}
