% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePeptideBindingLoad.R
\name{.predictBindingMHCnuggets}
\alias{.predictBindingMHCnuggets}
\title{MHCnuggets-based binding prediction}
\usage{
.predictBindingMHCnuggets(peptides, alleles)
}
\description{
Uses the predictMHCnuggets function for deep learning-based binding prediction.
MHCnuggets is available at \url{https://github.com/KarchinLab/mhcnuggets}.
}
\keyword{internal}
