% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePeptideBindingLoad.R
\name{.predictBindingNetMHCpan}
\alias{.predictBindingNetMHCpan}
\title{NetMHCpan-based binding prediction}
\usage{
.predictBindingNetMHCpan(peptides, alleles, backend_path)
}
\description{
NetMHCpan-based binding prediction
}
\keyword{internal}
