% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePeptideBindingLoad.R
\name{.predictBindingPWM}
\alias{.predictBindingPWM}
\title{PWM-based binding prediction (built-in)}
\usage{
.predictBindingPWM(peptides, alleles)
}
\description{
PWM-based binding prediction (built-in)
}
\keyword{internal}
