% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epletAUC.R
\name{epletAUC}
\alias{epletAUC}
\title{Calculate Eplet AUC Based on MFI}
\usage{
epletAUC(
  result_file,
  evidence_level = c("A1", "A2"),
  eplet_filter = 3,
  percPos_filter = 0.8,
  cut_min = 250,
  cut_max = 10000,
  cut_step = 250,
  plot_results = TRUE,
  palette = "inferno"
)
}
\arguments{
\item{result_file}{A data frame containing SAB results **or** a character 
string specifying the path to a SAB file in CSV, XLS, or XLSX format.}

\item{evidence_level}{A character string or vector of character strings
indicating the desired evidence levels to keep. Defaults to 
`c("A1", "A2")`, representing antibody-confirmed eplets. 
Other levels include: `B`, `D`, or `NULL` if no filter is desired.}

\item{eplet_filter}{An integer specifying the minimum number of times an 
eplet must appear in the assay before calculating the AUC. Defaults to `3`.}

\item{percPos_filter}{A numeric value between 0 and 1 representing the 
relative percent of positive beads (for a specific eplet) to use as a filter. 
Eplets below this threshold are excluded. Defaults to `0.8`.}

\item{cut_min}{An integer specifying the minimum MFI cutoff value. 
Defaults to `250`.}

\item{cut_max}{An integer specifying the maximum MFI cutoff value. 
Defaults to `10000`.}

\item{cut_step}{An integer specifying the increment step between the minimum 
and maximum MFI cutoff values. Defaults to `250`.}

\item{plot_results}{Logical. If `TRUE`, the function returns and prints a 
\code{ggplot} object illustrating the proportion of positive eplets at 
each cutoff. If `FALSE`, the function returns a summarized tibble.
Defaults to `TRUE`.}

\item{palette}{A character string indicating the color palette to use when 
plotting. Should be one of the palettes available through 
\link[grDevices]{hcl.pals} or a custom function. Defaults to `"inferno"`.}
}
\value{
If \code{plot_results = TRUE}, a \code{ggplot} object is returned. 
  If \code{plot_results = FALSE}, a tibble (data frame) is returned. If 
  \code{export_results = TRUE} and \code{plot_results = FALSE}, the results 
  are also written to a CSV file named \code{"epletAUC_results.csv"} in 
  the working directory.
}
\description{
This function reads in Single Antigen Bead (SAB) data from either a data frame
or a file path (CSV, XLS, or XLSX). It cleans and organizes the data, 
calculates how many eplets are positive above a range of MFI cutoffs,
computes the percentage of positive eplets, optionally filters them based on 
user-defined criteria, and finally calculates the area under the curve (AUC)
for each eplet. Depending on user arguments, it can either generate and print 
a ggplot **or** return/export the AUC-based results as a CSV.
}
\examples{
# Example 1: Plot data
epletAUC(
  result_file = deepMatchR_example[[1]],
  plot_results = TRUE, 
  percPos_filter = 0.9
)

# Example 2: Export data
result <- epletAUC(
  result_file = deepMatchR_example[[1]],
  plot_results = FALSE,
  export_results = TRUE, 
  percPos_filter = 0.9
)

}
