% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMatchR_SABModelparams.R
\docType{data}
\name{params}
\alias{params}
\title{SAB Model Parameters}
\format{
A list with two elements, one for each class.
}
\description{
A data object containing the transformation recipe for Class I
and Class II MFI values. This includes the method and for each allele, the
ceiling, mu (mean), and standard deviation values.
}
