% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_genotype.R
\name{print.hla_genotype}
\alias{print.hla_genotype}
\title{Print an hla_genotype object}
\usage{
\method{print}{hla_genotype}(x, ...)
}
\arguments{
\item{x}{An object of class `hla_genotype`.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the original object.
}
\description{
Print an hla_genotype object
}
\examples{
df <- data.frame(A_1 = "A*01:01", B_1 = "B*07:02")
geno <- hlaGeno(df)
print(geno)

}
