% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_genotype.R
\name{validateHlaGeno}
\alias{validateHlaGeno}
\title{Validate an hla_genotype object}
\usage{
validateHlaGeno(x)
}
\arguments{
\item{x}{An object to validate.}
}
\value{
`TRUE` if the object is a valid `hla_genotype` object, otherwise
  throws an error.
}
\description{
Validate an hla_genotype object
}
\keyword{internal}
