% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deepMatchR_cregs}
\alias{deepMatchR_cregs}
\title{CREG-Allele Mapping Data (Dummy)}
\format{
A data frame with 6 rows and 2 variables:
\describe{
  \item{allele}{An HLA allele string.}
  \item{creg}{The CREG group corresponding to the allele.}
}
}
\usage{
deepMatchR_cregs
}
\description{
A dummy dataset containing a mapping of HLA alleles to their
Cross-Reactive Groups (CREGs). This is a placeholder file to allow the
package to build and run correctly. The data is not comprehensive or
scientifically validated.
}
\keyword{datasets}
