% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepMatchR_example.R
\docType{data}
\name{deepMatchR_example}
\alias{deepMatchR_example}
\title{Simulated single antigen bead (SAB) assay
for class I and class II and PRA assay
with fixed formatting.}
\description{
Simulated single antigen bead (SAB) assay
for class I and class II and PRA assay
with fixed formatting.
}
