% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlleleSequence.R
\name{getAlleleSequence}
\alias{getAlleleSequence}
\title{Get Amino Acid Sequence for an HLA Allele}
\usage{
getAlleleSequence(allele_name)
}
\arguments{
\item{allele_name}{A character string representing the HLA allele name
(e.g., "A*01:01").}
}
\value{
A character string representing the amino acid sequence.
}
\description{
This function retrieves the amino acid sequence for a given HLA allele
name using the `immReferent` package.
}
\examples{
\dontrun{
# Get amino acid sequence for an HLA-A allele
seq <- getAlleleSequence("A*01:01")
nchar(seq)  # Length of sequence

# Get sequence for HLA-B allele
seq_b <- getAlleleSequence("B*07:02")
}

}
