% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlleleSequence.R
\name{getSequenceStats}
\alias{getSequenceStats}
\title{Get Sequence Statistics}
\usage{
getSequenceStats(sequences, type = c("PROT", "NUC"))
}
\arguments{
\item{sequences}{Named list of sequences (output from batchGetSequences)}

\item{type}{Type of sequences ("PROT" or "NUC")}
}
\value{
Data frame with sequence statistics
}
\description{
Analyzes sequences to provide statistics about length, composition, etc.
}
\examples{
alleles <- c("A*01:01", "A*02:01", "B*07:02")
seqs <- batchGetSequences(alleles)
stats <- getSequenceStats(seqs, type = "PROT")

}
