% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_genotype.R
\name{hlaGeno}
\alias{hlaGeno}
\title{Create an hla_genotype object}
\usage{
hlaGeno(df)
}
\arguments{
\item{df}{A data frame where rows are individuals and columns represent
HLA alleles (e.g., A_1, A_2, B_1, B_2, ...).}
}
\value{
An object of class `hla_genotype`.
}
\description{
Creates a new `hla_genotype` object from a data frame of HLA calls. The
object is a list containing the genotype data and a record of the loci
present.
}
\examples{
# Create a genotype from a data frame
recipient <- data.frame(
  A_1 = "A*01:01", A_2 = "A*02:01",
  B_1 = "B*07:02", B_2 = "B*08:01"
)
geno <- hlaGeno(recipient)
print(geno)

# Access components
geno$locus_present
geno$data

}
