% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyMismatch.R
\name{quantifyEpletMismatch}
\alias{quantifyEpletMismatch}
\title{Quantify Eplet Mismatches Between Two Alleles}
\usage{
quantifyEpletMismatch(allele1, allele2)
}
\arguments{
\item{allele1}{A character string for the first HLA allele (e.g., "A*01:01").}

\item{allele2}{A character string for the second HLA allele (e.g., "A*02:01").}
}
\value{
An integer representing the number of mismatched eplets.
}
\description{
This function calculates the number of eplet mismatches between two HLA
alleles. It uses the internal `deepMatchR_eplets` dataset.
}
