% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyMismatch.R
\name{quantifyMismatch}
\alias{quantifyMismatch}
\title{Quantify Amino Acid Mismatches Between Two Sequences}
\usage{
quantifyMismatch(sequence1, sequence2)
}
\arguments{
\item{sequence1}{A character string representing the first amino acid sequence.}

\item{sequence2}{A character string representing the second amino acid sequence.}
}
\value{
An integer representing the number of mismatches.
}
\description{
This function compares two amino acid sequences of the same length and
returns the number of positions at which the amino acids differ.
}
\examples{
seq1 <- "YFAMYGEKVAHTHVDTLYVRYHY"
seq2 <- "YFDMYGEKVAHTHVDTLYVRFHY"
quantifyMismatch(seq1, seq2)
#> [1] 2

}
